pyproject.toml
[project]
name = "academic-pretraining"
version = "1.0.0"
description = "Codebase for the paper '$100K or 100 Days: Trade-offs when Pre-Training with Academic Resources'"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Apoorv Khandelwal", email = "mail@apoorvkh.com" }]
dependencies = [
  # training libraries
  "accelerate==0.32.0",
  "deepspeed==0.14.4",
  "tabulate==0.9.0",      # for counting flops in pytorch
  "torch==2.3.1",
  "torchvision==0.18.1",
  "transformers==4.42.3",
  "triton==2.3.1",
  # experiment execution
  "ai2-tango==1.3.2",
  "torchrunx @ git+https://github.com/apoorvkh/torchrunx",
  "submitit==1.5.1",
  "tyro==0.8.10",
  # plotting
  "marimo==0.8.18",
  "polars[pandas]==1.7.1",
  "altair==5.4.1",
  "vl-convert-python==1.6.1", # altair[save]
  # linting, formatting, type checking
  "ruff==0.6.2",
  "pyright==1.1.377",
]

[project.optional-dependencies]
# custom kernels for mamba
mamba = [
  "causal-conv1d @ https://github.com/Dao-AILab/causal-conv1d/releases/download/v1.4.0/causal_conv1d-1.4.0+cu122torch2.3cxx11abiFALSE-cp311-cp311-linux_x86_64.whl",
  "mamba-ssm @ https://github.com/state-spaces/mamba/releases/download/v2.2.2/mamba_ssm-2.2.2+cu122torch2.3cxx11abiFALSE-cp311-cp311-linux_x86_64.whl",
]

[tool.pixi.project]
channels = ["nvidia/label/cuda-12.1.0", "conda-forge"]
platforms = ["linux-64"]

[tool.pixi.activation]
scripts = [".env"] # auto `source .env` in `pixi shell`

[tool.pixi.environments]
default = { features = ["mamba"], solve-group = "default" }

[tool.pixi.dependencies]
gcc = "11.2.0"
gxx = "11.2.0"
python = "3.11.9"
pip = "24.2"
uv = "0.3.3"
cuda = "12.1.0"

[tool.ruff]
line-length = 120
src = ["src", "experiments", "scripts"]
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E402"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"scripts/plotting/*.py" = [
  "E712", # Allow `== bool` in Polars
]

[tool.pyright]
include = ["src", "experiments", "scripts"]
exclude = ["scripts/plotting"]
pythonVersion = "3.11"
reportRedeclaration = "none"
